updated controller

+++ rewardsplus/src/main/java/com/company/rewardsplus/dao/CustomerInfoRepository.java
@@ -7,6 +7,6 @@ import com.company.rewardsplus.model.CustomerInfo;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface CustomerInfoRepository extends CrudRepository<CustomerInfo, Long> {
+public interface CustomerInfoRepository extends CrudRepository<CustomerInfo, Integer> {
 	
 }
